Move files to support cards page
