--- client/cards.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Board</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/board/main.jsx"></script>
-  </body>
-</html>
