+++ client/index.html
@@ -11,7 +11,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/index/main.jsx"></script>
+    <script type="module" src="/src/main.jsx"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
   </body>
 </html>
