+++ client/src/index/HomePage.jsx
@@ -1,54 +1,38 @@
 import { useState, useEffect } from "react";
-import "./App.css";
+import "./HomePage.css";
 import BrowseBar from "./components/BrowseBar";
 import BoardCard from "./components/BoardCard";
-import API from "../api";
+import propTypes from "prop-types";
 
-function App() {
-    const [userData, setUserData] = useState({
-        id: 1,
-        username: "Guest",
-        boardsCreated: [],
-        boardsLiked: [],
-        boardsDisliked: [],
-    });
-    const [boardData, setBoardData] = useState([]);
+HomePage.propTypes = {
+    userData: propTypes.array.isRequired,
+    boardData: propTypes.array.isRequired,
+}
 
+function HomePage(props) {
     const [category, setCategory] = useState("All");
-
-    // initialize the user/boards when loaded
-    useEffect(() => {
-        API.getUserData(setUserData, userData["id"]);
-        API.getBoardData(setBoardData);
-    }, []);
-
-    // whenever boardData changes, update user data as well
-    // as user data stores board creation/like/dislike data
-    useEffect(() => {
-        API.getUserData(setUserData, userData["id"]);
-    }, [boardData]);
 
     return (
         <>
             <header className="d-flex justify-content-between text-white bg-primary px-3 py-2">
                 <h2 className="m-0">Kudos Board</h2>
                 <button type="button" className="btn btn-outline-light">
-                    {userData["username"]}
+                    {props.userData[0]["username"]}
                 </button>
             </header>
             <BrowseBar
-                userData={[userData, setUserData]}
-                boardDataFunc={setBoardData}
+                userData={props.userData}
+                boardDataFunc={props.boardData[1]}
                 category={[category, setCategory]}
             />
 
             <div className="row row-cols-2 row-cols-md-4 row-cols-lg-5 justify-content-center w-100">
-                {boardData.map((board, index) => {
+                {props.boardData[0].map((board, index) => {
                     return (
                         <BoardCard
                             key={index}
                             board={board}
-                            isOwned={userData["boardsCreated"].some(
+                            isOwned={props.userData[0]["boardsCreated"].some(
                                 (createdBoard) => createdBoard["id"] === board["id"]
                             )}
                         />
@@ -65,4 +49,4 @@ function App() {
     );
 }
 
-export default App;
+export default HomePage;
